+++ src/main/java/pl/com/bottega/documentmanagement/api/DocumentCriteria.java
@@ -1,4 +1,4 @@
-package pl.com.bottega.documentmanagement.domain;
+package pl.com.bottega.documentmanagement.api;
 
 /**
  * Created by maciuch on 12.06.16.
