+++ src/main/java/pl/com/bottega/documentmanagement/api/DocumentsCatalog.java
@@ -1,10 +1,6 @@
 package pl.com.bottega.documentmanagement.api;
 
-import com.google.common.collect.Iterables;
-import pl.com.bottega.documentmanagement.domain.Document;
-import pl.com.bottega.documentmanagement.domain.DocumentCriteria;
 import pl.com.bottega.documentmanagement.domain.DocumentNumber;
-import pl.com.bottega.documentmanagement.domain.repositories.DocumentRepository;
 
 import static com.google.common.base.Preconditions.checkNotNull;
