+++ src/main/java/pl/com/bottega/documentmanagement/domain/repositories/DocumentRepository.java
@@ -1,6 +1,5 @@
 package pl.com.bottega.documentmanagement.domain.repositories;
 
-import pl.com.bottega.documentmanagement.domain.DocumentCriteria;
 import pl.com.bottega.documentmanagement.domain.Document;
 import pl.com.bottega.documentmanagement.domain.DocumentNumber;
 
@@ -12,7 +11,5 @@ public interface DocumentRepository {
     void save(Document document);
 
     Document load(DocumentNumber documentNumber);
-
-    Iterable<Document> find(DocumentCriteria criteria);
 
 }
